refactor: better tail recursion support
